保存一下，避免被覆盖掉

+++ backend/src/main/java/com/wyl/backend/controller/user/UserController.java
@@ -59,7 +59,7 @@ public class UserController {
      * @return
      */
     @PostMapping("/login")
-    public int login(@RequestBody UserInfo userInfo) {
+    public long login(@RequestBody UserInfo userInfo) {
         try {
             List<UserInfo> userinfo = userMapper.select();
             for (UserInfo u : userinfo) {
